requirements.txt
streamlit
tableone
pandas
numpy
matplotlib
seaborn
scipy
plotly-express
statsmodels
pickle